+++ frontend/src/components/Pages/Proprietaire.vue
@@ -167,15 +167,17 @@ export default {
         },
 
         DeleteProprietaire(id){
-            const options = {
-            url: 'http://127.0.0.1:8000/api/v1/deleteProprietaire/'+id,
-            method: 'DELETE',
-            };
+            if (confirm("Are You sure you want to delete this Parcelle") == true) {
+				const options = {
+                    url: 'http://127.0.0.1:8000/api/v1/deleteProprietaire/'+id,
+                    method: 'DELETE',
+                };
 
-            axios(options)
-            .then(response => {
-                this.getProprietaires();
-            });
+                axios(options)
+                .then(response => {
+                    this.getProprietaires();
+                });
+			}
         },
