+++ frontend/src/components/Pages/Village.vue
@@ -138,15 +138,17 @@ export default {
         },
 
         DeleteVillage(id){
-            const options = {
-            url: 'http://127.0.0.1:8000/api/v1/deleteVillage/'+id,
-            method: 'DELETE',
-            };
+            if (confirm("Are You sure you want to delete this Parcelle") == true) {
+				const options = {
+                    url: 'http://127.0.0.1:8000/api/v1/deleteVillage/'+id,
+                    method: 'DELETE',
+                };
 
-            axios(options)
-            .then(response => {
-                this.getVillages();
-            });
+                axios(options)
+                .then(response => {
+                    this.getVillages();
+                });
+			}
         },
